test: merge coverage improvements into existing test files
