+++ domain_test.go
@@ -198,3 +198,157 @@ func TestDomain_GetWithVar(t *testing.T) {
 		t.Errorf("Expected 'MyText' but got '%s'", tr)
 	}
 }
+
+func TestDomain_Append(t *testing.T) {
+	d := NewDomain()
+	d.Set("test", "translated")
+
+	b := []byte("prefix: ")
+	res := d.Append(b, "test")
+	if string(res) != "prefix: translated" {
+		t.Errorf("Expected 'prefix: translated', got '%s'", string(res))
+	}
+
+	res = d.Append(nil, "missing")
+	if string(res) != "missing" {
+		t.Errorf("Expected 'missing', got '%s'", string(res))
+	}
+}
+
+func TestDomain_AppendN(t *testing.T) {
+	d := NewDomain()
+	d.SetN("one", "many", 1, "singular")
+	d.SetN("one", "many", 2, "plural")
+
+	res := d.AppendN(nil, "one", "many", 1)
+	if string(res) != "singular" {
+		t.Errorf("Expected 'singular', got '%s'", string(res))
+	}
+
+	res = d.AppendN(nil, "one", "many", 2)
+	if string(res) != "plural" {
+		t.Errorf("Expected 'plural', got '%s'", string(res))
+	}
+
+	res = d.AppendN(nil, "missing", "missings", 1)
+	if string(res) != "missing" {
+		t.Errorf("Expected 'missing', got '%s'", string(res))
+	}
+
+	res = d.AppendN(nil, "missing", "missings", 2)
+	if string(res) != "missings" {
+		t.Errorf("Expected 'missings', got '%s'", string(res))
+	}
+}
+
+func TestDomain_AppendC(t *testing.T) {
+	d := NewDomain()
+	d.SetC("test", "ctx", "translated")
+
+	res := d.AppendC(nil, "test", "ctx")
+	if string(res) != "translated" {
+		t.Errorf("Expected 'translated', got '%s'", string(res))
+	}
+
+	res = d.AppendC(nil, "test", "wrong_ctx")
+	if string(res) != "test" {
+		t.Errorf("Expected 'test', got '%s'", string(res))
+	}
+}
+
+func TestDomain_AppendNC(t *testing.T) {
+	d := NewDomain()
+	d.SetNC("one", "many", "ctx", 1, "singular")
+	d.SetNC("one", "many", "ctx", 2, "plural")
+
+	res := d.AppendNC(nil, "one", "many", 1, "ctx")
+	if string(res) != "singular" {
+		t.Errorf("Expected 'singular', got '%s'", string(res))
+	}
+
+	res = d.AppendNC(nil, "one", "many", 2, "ctx")
+	if string(res) != "plural" {
+		t.Errorf("Expected 'plural', got '%s'", string(res))
+	}
+}
+
+func TestDomain_SetNC(t *testing.T) {
+	d := NewDomain()
+	d.SetNC("one", "many", "ctx", 1, "singular")
+	// Update existing
+	d.SetNC("one", "many", "ctx", 1, "singular_updated")
+
+	res := d.GetNC("one", "many", 1, "ctx")
+	if res != "singular_updated" {
+		t.Errorf("Expected 'singular_updated', got '%s'", res)
+	}
+
+	// New one in existing context
+	d.SetNC("two", "plural_two", "ctx", 1, "two_singular")
+	res = d.GetNC("two", "plural_two", 1, "ctx")
+	if res != "two_singular" {
+		t.Errorf("Expected 'two_singular', got '%s'", res)
+	}
+}
+
+func TestDomain_Refs(t *testing.T) {
+	d := NewDomain()
+	refs := []string{"file.go:10", "file.go:20"}
+	d.SetRefs("test", refs)
+
+	gotRefs := d.GetRefs("test")
+	if len(gotRefs) != 2 || gotRefs[0] != refs[0] || gotRefs[1] != refs[1] {
+		t.Errorf("Expected %v, got %v", refs, gotRefs)
+	}
+
+	if d.GetRefs("missing") != nil {
+		t.Error("Expected nil for missing refs")
+	}
+
+	// Update refs
+	newRefs := []string{"file.go:30"}
+	d.SetRefs("test", newRefs)
+	gotRefs = d.GetRefs("test")
+	if len(gotRefs) != 1 || gotRefs[0] != newRefs[0] {
+		t.Errorf("Expected %v, got %v", newRefs, gotRefs)
+	}
+}
+
+func TestDomain_HeaderMap(t *testing.T) {
+	d := NewDomain()
+	d.Headers.Del("Missing") // No-op but increases coverage
+
+	d.Headers.Set("Key", "Value")
+	if d.Headers.Get("Key") != "Value" {
+		t.Error("Header Get/Set failed")
+	}
+
+	d.Headers.Add("Key", "Value2")
+	values := d.Headers.Values("Key")
+	if len(values) != 2 || values[1] != "Value2" {
+		t.Error("Header Add failed")
+	}
+
+	d.Headers.Del("Key")
+	if d.Headers.Get("Key") != "" {
+		t.Error("Header Del failed")
+	}
+
+	var nilHeaders HeaderMap
+	if nilHeaders.Get("Any") != "" {
+		t.Error("Nil headers should return empty string")
+	}
+	if nilHeaders.Values("Any") != nil {
+		t.Error("Nil headers should return nil values")
+	}
+}
+
+func TestDomain_SetPluralResolver(t *testing.T) {
+	d := NewDomain()
+	d.SetPluralResolver(func(n int) int {
+		return 5
+	})
+	if d.pluralForm(10) != 5 {
+		t.Error("Custom plural resolver failed")
+	}
+}
